requirements.txt
adal==1.2.4
azure-cognitiveservices-vision-computervision==0.6.0
azure-common==1.1.25
azure-graphrbac==0.61.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==10.2.0
azure-mgmt-storage==11.2.0
azureml==0.2.7
azureml-core==1.13.0
backports.tempfile==1.0
backports.weakref==1.0.post1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
contextlib2==0.6.0.post1
convertapi==1.4.0
cryptography==3.1.1
docker==4.3.1
filelock==3.0.12
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
importlib-metadata==2.0.0
isodate==0.6.0
itsdangerous==1.1.0
jeepney==0.4.3
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonpickle==1.4.1
MarkupSafe==1.1.1
msrest==0.6.19
msrestazure==0.6.4
ndg-httpsclient==0.5.1
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.3
pathspec==0.8.0
Pillow==7.2.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.23.2
scipy==1.5.4
SecretStorage==3.1.2
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.20
threadpoolctl==2.1.0
urllib3==1.25.10
websocket-client==0.57.0
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.3.1
